refactor(std-client): move setup utils from setupMUDNetwork to utils
