+++ packages/std-client/src/setup/index.ts
@@ -1,2 +1,2 @@
 export { setupMUDNetwork } from "./setupMUDNetwork";
-export type { SetupContractConfig } from "./setupMUDNetwork";
+export type { SetupContractConfig } from "./types";
